requirements.txt
dask_ml==0.12.0
ase==3.17.0
scipy==1.1.0
numpy==1.16.2
msgpack_numpy==0.4.4.2
matplotlib==3.0.2
torch==1.0.1.post2
setuptools==40.8.0
msgpack_python==0.5.6
scikit_learn==0.20.3
